+++ CHANGELOG.md
@@ -8,6 +8,8 @@ All notable changes to this project will be documented in this file.
   the put() and get() behaviors from the database so that each
   datasetType knows how to handle its own data.
 - Added `__version__` field to all modules.
+- MergeFangTS stats computer for computing statistics on merged
+  localizations in Fang's ThunderSTORM column format.
 
 ### Changed
 - The `particle` column is now saved when using the MergeFang stats
@@ -15,6 +17,9 @@ All notable changes to this project will be documented in this file.
   it was not being saved because the the stats computer was making it
   an index of the output DataFrame; the CSVBatchProcessor does not
   save DataFrame indexes.
+- Merge processor now accepts a coordinate column name parameter for
+  merging columns with custom names.
+- Merge processor attributes are now public.
 
 ## [v0.2.0]
 ### Added